feat: move anon fn, let, and expect
